requirements.txt
requests>=2.28.0
pyyaml>=6.0
flask>=3.0.0
